refactor(utils): 统一 utils 目录文件命名为 kebab-case 格式
- 为什么改：utils 目录文件命名（camelCase）与项目其他模块（handlers、middlewares、services）的 kebab-case
命名规范不一致，影响代码库整体一致性
- 改了什么：
  1. 重命名 3 个工具文件：PathUtils.ts → path-utils.ts、TypeFieldNormalizer.ts →
type-field-normalizer.ts、VersionUtils.ts → version-utils.ts
  2. 删除未使用的重复文件：mcpServerUtils.ts（类型已在 shared-types 包中定义）
  3. 重命名测试文件：TypeFieldNormalizer.test.ts → type-field-normalizer.test.ts
  4. 新增统一导出文件：utils/index.ts，提供所有工具类的集中导出接口
  5. 更新 7 个文件的导入引用，使用新的 kebab-case 文件名
- 影响范围：
  - 受影响模块：lib/mcp、handlers、utils
  - 向后兼容：类名/命名空间名保持 PascalCase 不变（如 PathUtils、TypeFieldNormalizer、VersionUtils），仅文件名变更
  - 导入方式：现有代码需更新导入路径为 kebab-case 格式
- 验证方式：
  1. TypeScript 类型检查通过（pnpm check:type）
  2. 代码规范检查通过（pnpm lint）
  3. 完整质量检查通过（pnpm check:all）
  4. 所有测试用例通过（567 tests）
