+++ apps/backend/handlers/__tests__/mcp-manage.handler.test.ts
@@ -2,7 +2,7 @@ import path from "node:path";
 import type { MCPServiceManager } from "@/lib/mcp";
 import { MCPErrorCode } from "@errors/MCPErrors.js";
 import type { EventBus } from "@services/event-bus.service.js";
-import { TypeFieldNormalizer } from "@utils/TypeFieldNormalizer.js";
+import { TypeFieldNormalizer } from "@utils/type-field-normalizer.js";
 import type { ConfigManager } from "@xiaozhi-client/config";
 import type { MCPServerConfig } from "@xiaozhi-client/config";
 import type { Context } from "hono";
