+++ apps/backend/handlers/mcp-manage.handler.ts
@@ -10,7 +10,7 @@ import {
   MCPErrorCode,
 } from "@root/errors/MCPErrors.js";
 import { getEventBus } from "@services/event-bus.service.js";
-import { TypeFieldNormalizer } from "@utils/TypeFieldNormalizer.js";
+import { TypeFieldNormalizer } from "@utils/type-field-normalizer.js";
 import type { ConfigManager, MCPServerConfig } from "@xiaozhi-client/config";
 import { normalizeServiceConfig } from "@xiaozhi-client/config";
 import type { Context } from "hono";
