+++ apps/backend/utils/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Utils 统一导出模块
+ * 提供 utils 目录下所有工具类的统一导出接口
+ */
+
+export { PathUtils } from "./path-utils.js";
+export { TypeFieldNormalizer } from "./type-field-normalizer.js";
+export { VersionUtils } from "./version-utils.js";
