+++ apps/backend/utils/path-utils.ts
