+++ apps/backend/utils/type-field-normalizer.ts
